+++ backend/src/main/java/com/net1707/backend/model/OrderDetail.java
@@ -3,6 +3,8 @@ package com.net1707.backend.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
@@ -14,7 +16,7 @@ public class OrderDetail {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer orderDetailId;
+    private Long orderDetailId;
 
     @ManyToOne
     @JoinColumn(name = "orderId", nullable = false)
@@ -28,5 +30,5 @@ public class OrderDetail {
     private Integer quantity;
 
     @Column(nullable = false)
-    private Float unitPrice;
+    private BigDecimal unitPrice;
 }
